+++ src/r3_core/src/time/time.rs
@@ -2,7 +2,7 @@ use core::{fmt, ops};
 
 use crate::{
     time::Duration,
-    utils::{Init, ZeroInit, Zeroable},
+    utils::{Init, Zeroable},
 };
 
 /// Represents a timestamp used by the API surface of R3-OS.
@@ -22,10 +22,6 @@ pub struct Time {
 impl Init for Time {
     const INIT: Self = Self::ZERO;
 }
-
-// Safety: `Time` is `repr(transparent)` and the only inner field is `u64`,
-//         which is `ZeroInit`
-unsafe impl ZeroInit for Time {}
 
 impl Default for Time {
     fn default() -> Self {